added support for default iframe with contentURL and styling
