+++ packages/quantic/force-app/main/default/lwc/quanticQuickviewContent/quanticQuickviewContent.css
@@ -1,3 +1,12 @@
 @import 'c/quanticResultQuickview';
 @import 'c/quanticResultActionStyles';
 
+.iframe-wrapper {
+  width: 100%;
+  height: 100%;
+  display: flex;
+}
+
+.iframe-wrapper iframe {
+  border: none;
+}
