+++ packages/quantic/force-app/main/default/lwc/quanticQuickviewContent/quanticQuickviewDefault.css
