+++ packages/quantic/force-app/main/default/lwc/quanticQuickviewContent/quanticQuickviewDefault.html
@@ -1,10 +1,12 @@
 <template>
-  <!-- <iframe
-      id="quickview__content-container"
-      class="quickview__content-container slds-modal__content slds-p-around_large slds-wrap"
-      src={contentURL}
-      height="100%"
-      onload={onIframeLoaded}
-  ></iframe> -->
-  <div>hello</div>
+  <div class="iframe-wrapper">
+    <iframe
+        id="quickview__content-container"
+        class="quickview__content-container slds-modal__content slds-p-around_large slds-wrap"
+        src={contentURI}
+        height="100%"
+        width="100%"
+        onload={onIframeLoaded}
+    ></iframe>
+  </div>
 </template>
