+++ packages/quantic/force-app/main/default/lwc/quanticQuickviewContent/quanticQuickviewYoutube.html
@@ -8,7 +8,7 @@
     allowfullscreen
     referrerpolicy="same-origin"
     onload={handleIframeLoaded}
-    sandbox="allow-same-origin allow-scripts allow-popups allow-top-navigation allow-top-navigation-by-user-activation"
+    sandbox="allow-same-origin allow-scripts allow-popups allow-top-navigation allow-top-navigation-by-user-activation allow-presentation"
     src={youtubeURL}
     ></iframe>
   </div>
