+++ packages/quantic/force-app/main/default/lwc/quanticResultQuickview/quanticResultQuickview.html
@@ -92,14 +92,20 @@
             ></lightning-spinner>
           </div>
         </template> -->
+
         <c-quantic-quickview-content
           id="quickview__content-container"
           class="quickview__content-container slds-modal__content slds-wrap"
           engine-id={engineId}
           result={result}
+          content-url={contentURL}
           height="100%"
           is-loading={isLoading}
         ></c-quantic-quickview-content>
+
+        <!-- <iframe id="quickview__content-container"
+          class="quickview__content-container slds-modal__content slds-p-around_large slds-wrap" src={contentURL}
+          height="100%" onload={onIframeLoaded}></iframe> -->
 
         <template if:false={isLoading}>
           <footer
